refactor: create custom hook useChatHandler
